add fibonacci

+++ src/com/javatasks/fibonacci/Fibonacci.java
@@ -11,7 +11,7 @@ public class Fibonacci {
         System.out.println("Цикл " + fib2(a));
     }
 
-    static long fib(int n) {
+    static int fib(int n) {
         if (n <= 1) {
             return n;
         } else if (n >= 49) {
@@ -20,14 +20,14 @@ public class Fibonacci {
         return fib(n - 1) + fib(n - 2);
     }
 
-    static long fib2(int n) {
+    static int fib2(int n) {
         if(n >= 49){
             System.out.println("Введите число мменьше 48");
 
         }
-        long first = 0;
-        long second = 1;
-        long result = n;
+        int first = 0;
+        int second = 1;
+        int result = n;
         for (int i = 1; i < n; i++) {
             result = first + second;
             first = second;
